実装するクラスを間違えていたので修正
